* improve RLE statistics (bugs + 64-bit counters)
* fix bugs detected during integration tests
* accept jump of sequence number to 0 to manage modem reconnection

+++ tests/test_non_regression.c
@@ -716,32 +716,32 @@ static int test_encap_and_decap(const char *const src_filename)
 			u_int8_t frag_id;
 			for (frag_id = 0; frag_id < 8; ++frag_id) {
 				printf("===\tFrag ID %u\n", frag_id);
-				printf("===\ttransmitter in:             %zu\n",
-				       rle_transmitter_stats_get_counter_sdus_in(transmitter, frag_id));
-				printf("===\ttransmitter sent:           %zu\n",
-				       rle_transmitter_stats_get_counter_sdus_sent(transmitter, frag_id));
-				printf("===\ttransmitter dropped:        %zu\n",
-				       rle_transmitter_stats_get_counter_sdus_dropped(transmitter, frag_id));
-				printf("===\ttransmitter bytes in:       %zu\n",
-				       rle_transmitter_stats_get_counter_bytes_in(transmitter, frag_id));
-				printf("===\ttransmitter bytes sent:     %zu\n",
-				       rle_transmitter_stats_get_counter_bytes_sent(transmitter, frag_id));
-				printf("===\ttransmitter bytes dropped:  %zu\n",
-				       rle_transmitter_stats_get_counter_bytes_dropped(transmitter, frag_id));
-				printf("===\treceiver received:          %zu\n",
-				       rle_receiver_stats_get_counter_sdus_received(receiver, frag_id));
-				printf("===\treceiver reassembled:       %zu\n",
-				       rle_receiver_stats_get_counter_sdus_reassembled(receiver, frag_id));
-				printf("===\treceiver lost:              %zu\n",
-				       rle_receiver_stats_get_counter_sdus_lost(receiver, frag_id));
-				printf("===\treceiver dropped:           %zu\n",
-				       rle_receiver_stats_get_counter_sdus_dropped(receiver, frag_id));
-				printf("===\treceiver bytes received:    %zu\n",
-				       rle_receiver_stats_get_counter_bytes_received(receiver, frag_id));
-				printf("===\treceiver bytes reassembled: %zu\n",
-				       rle_receiver_stats_get_counter_bytes_reassembled(receiver, frag_id));
-				printf("===\treceiver bytes dropped:     %zu\n",
-				       rle_receiver_stats_get_counter_bytes_dropped(receiver, frag_id));
+				printf("===\ttransmitter in:             %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_sdus_in(transmitter, frag_id));
+				printf("===\ttransmitter sent:           %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_sdus_sent(transmitter, frag_id));
+				printf("===\ttransmitter dropped:        %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_sdus_dropped(transmitter, frag_id));
+				printf("===\ttransmitter bytes in:       %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_bytes_in(transmitter, frag_id));
+				printf("===\ttransmitter bytes sent:     %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_bytes_sent(transmitter, frag_id));
+				printf("===\ttransmitter bytes dropped:  %llu\n",
+						(long long unsigned int)rle_transmitter_stats_get_counter_bytes_dropped(transmitter, frag_id));
+				printf("===\treceiver received:          %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_received(receiver, frag_id));
+				printf("===\treceiver reassembled:       %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_reassembled(receiver, frag_id));
+				printf("===\treceiver lost:              %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_lost(receiver, frag_id));
+				printf("===\treceiver dropped:           %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_sdus_dropped(receiver, frag_id));
+				printf("===\treceiver bytes received:    %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_received(receiver, frag_id));
+				printf("===\treceiver bytes reassembled: %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_reassembled(receiver, frag_id));
+				printf("===\treceiver bytes dropped:     %llu\n",
+						(long long unsigned int)rle_receiver_stats_get_counter_bytes_dropped(receiver, frag_id));
 				printf("\n");
 			}
 		}
